Remove superfluous EchoServer class
This was added in https://github.com/JetBrains/attachme/commit/5d97e397514e4393bbeae641ba5c3ed008c47e8b,
but I think it was an earlier experiment that never ended up being used
(there's a separate `EchoService` and `Client` class in the `example/`
class which was added in that commit). The problem with including this
in `agent/src/main/java/` is that it gets included in the published
`.jar` file, which is almost certainly unintentional.